Move playlist init to hooks
Next pre-renders state init if its just a function call and not a useEffect
This hits browser APIs, so no-go for initial state render
